MyPython_test/streamlit/requirements.txt
altair==4.2.2
altgraph==0.17.2
ansicon==1.89.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blessed==1.19.1
blinker==1.5
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.2.2
flyingcircus==0.1.4.0
fonttools==4.33.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
h11==0.13.0
idna==3.3
importlib-metadata==6.0.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jinxed==1.2.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
MouseInfo==0.1.3
mutagen==1.45.1
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
pikepdf==6.2.7
Pillow==9.1.1
plyer==2.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pyarrow==11.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodomex==3.14.1
pydeck==0.8.0
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
Pympler==1.0.1
PyMsgBox==1.0.9
pyocr==0.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
pytk==0.0.2.1
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.5
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.3.1
selenium==4.1.5
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==6.4.2
six==1.16.0
slack-bolt==1.14.0
slack-sdk==3.17.1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
stack-data==0.2.0
streamlit==1.18.1
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
traitlets==5.2.1.post0
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.10.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
watchdog==2.2.1
wcwidth==0.2.5
webdriver-manager==3.5.4
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wsproto==1.1.0
youtube-dl==2021.12.17
yt-dlp==2022.5.18
zipp==3.13.0
